+++ scripts/build-utils.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 const fs = require("fs");
 const path = require("path");
-const { publishedComponents, publishedTSComponents } = require("../webpack/published-components");
+const { publishedTSComponents, publishedStorybookComponents } = require("../webpack/published-components");
 
 function createFoldersIfNotExist() {
   // if dist is not exist let's create it
@@ -15,6 +15,12 @@ function createFoldersIfNotExist() {
   const iconsDir = path.join(distDir, `/icons`);
   if (!fs.existsSync(iconsDir)) {
     fs.mkdirSync(iconsDir);
+  }
+
+  // if dist/storybook does not exist let's create it
+  const storybookDir = path.join(distDir, `/storybook`);
+  if (!fs.existsSync(storybookDir)) {
+    fs.mkdirSync(storybookDir);
   }
 }
 
@@ -35,7 +41,16 @@ function buildComponentsTypesIndexFile() {
   convertExportsToFile(exports, "types.d.ts");
 }
 
+function buildStorybookComponentsIndexFile() {
+  const exports = Object.entries(publishedStorybookComponents).map(([name, _path]) => {
+    const fileName = name.split("/").slice(-1);
+    return buildComponentExport(fileName, `./${fileName}`);
+  });
+  convertExportsToFile(exports, "storybook/index.js");
+}
+
 module.exports = {
   createFoldersIfNotExist,
-  buildComponentsTypesIndexFile
+  buildComponentsTypesIndexFile,
+  buildStorybookComponentsIndexFile
 };