+++ src/storybook/components/component-description/component-description.jsx
@@ -16,3 +16,5 @@ ComponentDescription.defaultProps = {
   description: "",
   children: null
 };
+
+export default ComponentDescription;