+++ src/storybook/components/information-box-title/information-box-title.jsx
@@ -16,3 +16,5 @@ InformationBoxTitle.propTypes = {
 InformationBoxTitle.defaultProps = {
   component: ""
 };
+
+export default InformationBoxTitle;