+++ src/storybook/functions/createComponentTemplate.js
@@ -1,0 +1,8 @@
+import React from "react";
+
+export function createComponentTemplate(ComponentClass) {
+  // eslint-disable-next-line react/display-name
+  return args => <ComponentClass {...args} />;
+}
+
+export default createComponentTemplate;